+++ lib/src/extensions/double_operators.dart
@@ -14,24 +14,9 @@ extension DoubleOperators on List<double> {
   List<double> operator *(num scalar) {
     return List<double>.generate(length, (i) => this[i] * scalar);
   }
-
-  /// Returns `true` if the equality
-  /// `this(i) == other(i)` holds for each index `i`.
-  bool match(List<double> other, {double precision = 1e-12}) {
-    if (this == other) return true;
-    if (length != other.length) return false;
-    final it = iterator;
-    final oit = other.iterator;
-    while (it.moveNext() && oit.moveNext()) {
-      if ((it.current - oit.current).abs() > precision) {
-        return false;
-      }
-    }
-    return true;
-  }
 }
 
-extension DoubleIterableOperators on Iterable<double> {
+extension DoubleIterableMethods on Iterable<double> {
   /// Returns the sum of the entries.
   ///
   /// The iterable must not be empty.